Added Mod Menu as a suggestion in fabric.mod.json
REFACTOR: Mod Menu is now in the "suggested" section of fabric.mod.json

I also tried using YACL, as it seemed interesting, but I've decided against it for now. If the config gets too big, however, I'll consider making the switch to it, as it seems like it can be a lot more organized than MidnightLib.