0.0.2 - Attempted multiplayer support
FEAT: Changed the mixin target point for PlayerEntityMixin.java for the non-critical hit particles, in an attempt to make it work on multiplayer.
FEAT: Added new logger saying that particles are to be summoned.
REFACTOR: Disabled temporary loggers in StarParticle.java to help reduce logger spam, even with the spamLog config option enabled.