Bumped version once more.
REFACTOR: Bumped version to 0.0.7
REFACTOR: Changed PlayerEntityMixin.java's mixin from "private void" to void.

Something different happened this time... I don't think it showed the particle because it crashed on multiplayer. I'm using Visuality as a reference for this part, as if this works, I was going in the wrong direction this ENTIRE time.