Completely changed the method on how particles are spawned.
REFACTOR: A new mixin is being used in LivingEntity to know when to spawn the particles. This method works in multiplayer. The downside is that there are now inconsistencies when spawning particles.
REFACTOR: Used a new method to spawn the particles, meaning that they only spawn client-side. The downside to this is that I lose the "speed" variable, which was used for the particles' movement.
REFACTOR: Changed the "Particle Amount" and "Secondary Particle Amount" strings to "Rough Particle Amount" and "Rough Secondary Particle Amount"
FEAT: Added a new tooltip mentioning how particle amounts will sometimes be off.