New methods for calling when particles to be spawned. Complete refactor.
REFACTOR: Changed any method involved with spawning particles to calling the particles to be spawned.
REFACTOR: New method called "spawnParticles(boolean primary, ParticleEffect particle)", Allows both the primary and secondary methods to call this one method, using the switch function inside the call methods. Sadly, it seems like it might not be possible to make the mod spawn only ONE particle whenever the config is set to one, due to how Minecraft handles things.
REFACTOR: Added a note saying that if the particleAmount slider is set to 1, it will most likely not spawn one.