Config changes!
FEAT: New config option: Extra particles on death. This is for both the primary and secondary particles.
REFACTOR: Changed wording on quite a bit of stuff.
FIX: Finally fixed a bug where the particle amount would not be exact.
REMOVED: Removed the config option: Only on crit, as it is no longer possible to my knowledge.