Updated to new method and removed config options
REFACTOR: Updated all the code to work with the new method. Players will now activate the particles in single player, but I have yet to test it in multiplayer.
REFACTOR: Removed the config options which allowed the particles to only spawn if the damage source was a player. This was removed as it's no longer (easily) possible with the new method

Let's really hope this works in multiplayer once more. If it does, then that means I only have a few things left to finish up before release!